Used new prettify functions to format download info instead of hardcoding in place.
git-svn-id: https://192.168.3.139:8443/svn/Medical/trunk@5625 956a8c8d-885e-cb46-a0b9-a2f355b4c364

+++ Standalone/Controller/Download/Download.cs
@@ -99,7 +99,7 @@ namespace Medical
         /// <summary>
         /// The download speed in kb per second.
         /// </summary>
-        public double DownloadSpeed { get; set; }
+        public long DownloadSpeed { get; set; }
 
         public String StatusString { get; set; }
